Add goclean.sh and address linters.
Enable goclean.sh in travis

install gometalinter in Travis

+++ rpcutils/rpcclient.go
@@ -102,7 +102,7 @@ func BuildBlockHeaderVerbose(header *wire.BlockHeader, params *chaincfg.Params,
 
 	blockHeaderResult := dcrjson.GetBlockHeaderVerboseResult{
 		Hash:          header.BlockHash().String(),
-		Confirmations: int64(currentHeight - int64(header.Height)),
+		Confirmations: currentHeight - int64(header.Height),
 		Version:       header.Version,
 		PreviousHash:  header.PrevBlock.String(),
 		MerkleRoot:    header.MerkleRoot.String(),
